% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/importAllSnowData.R
\name{importAllSnowdata}
\alias{importAllSnowdata}
\title{Import Canadian snow station information data from .dly file}
\usage{
importAllSnowdata(fileLoc)
}
\arguments{
\item{fileLoc}{File path to .dly data}
}
\value{
nicely organized dataframe of snow data
}
\description{
Import Canadian snow station information data from .dly file
}
