% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/importAllStationdata.R
\name{importAllStationdata}
\alias{importAllStationdata}
\title{Import Canadian snow station information data from .dly file}
\usage{
importAllStationdata(fileLoc)
}
\arguments{
\item{fileLoc}{File path to .dly data}
}
\value{
nicely organized dataframe of snow data. ID is station ID, Name is station name, Lat is station latitude, Lon is station longitude, Elev is station elevation, Sdate is start date of observations (YYYYMM), Edate is end date of observations (YYYYMM), Nobs is number of observations
}
\description{
Import Canadian snow station information data from .dly file
}
