% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/importAllSnowData.R
\name{importAllSnowdata}
\alias{importAllSnowdata}
\title{Import Canadian snow station information data from .dly file}
\usage{
importAllSnowdata(fileLoc)
}
\arguments{
\item{fileLoc}{File path to .dly data}
}
\value{
nicely organized dataframe of snow data. ID is station ID, Year is the year of observation, Month is the month of observation, Day is the day of observation, SnowDepth is the depth of snow on the ground measured in centimeters, QC is a quality control flag where 0 is a valid observation and 999 is missing where the numbers inbetween indicate how accurate the estimate is (smaller is more accurate), CA is Climate archive flag where C= snow depth manually changed based on nearby stations | E= estimated | M=missing | T= trace (<.5 cm) | Z= value set to missing or filling of missing zero snow depths from snow cover
}
\description{
Import Canadian snow station information data from .dly file
}
