% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/importAllStationdata.R
\name{importAllStationdata}
\alias{importAllStationdata}
\title{Import Canadian snow station information data from .dly file}
\usage{
importAllStationdata(fileLoc)
}
\arguments{
\item{fileLoc}{File path to .dly data}
}
\value{
nicely organized dataframe of snow data
}
\description{
Import Canadian snow station information data from .dly file
}
